Improving list functionality for current_index

+++ custom_stream_api/lists/lists.py
@@ -53,28 +53,47 @@ def get_list(name):
         return []
 
 
-def get_list_item(list_name, index=None):
+def represents_int(s):
+    try:
+        int(s)
+    except ValueError:
+        return False
+    else:
+        return True
+
+
+def get_list_item(list_name, index):
+    """Index is 1-indexed, string random or next"""
     found_list = db.session.query(List).filter_by(name=list_name).one_or_none()
     if not found_list:
         raise Exception("List not found")
-    items = sorted(found_list.items, key=lambda list_item: list_item.id)
-    if len(items) == 0:
+    if len(found_list.items) == 0:
         raise Exception("Empty list")
+    items = sorted(found_list.items, key=lambda list_item: list_item.id)
+
     if index is None:
+        raise Exception("index not provided")
+    elif isinstance(index, int) or (isinstance(index, str) and represents_int(index)):
+        index = int(index)
+        if index > len(items):
+            raise Exception("Index too high")
+        elif index == 0:
+            raise Exception("Lists are 1-indexed.")
+    elif isinstance(index, str) and index.lower() == "random":
         index = random.choice(range(1, len(items) + 1))
+    elif isinstance(index, str) and index.lower() == "next":
+        # current_index is 0-indexed, adding 1 to be in sync with the rest
+        index = (found_list.current_index + 1) % len(items)
+        found_list.current_index = index
+        index += 1
     else:
-        if not isinstance(index, int):
-            if not index.isdigit():
-                return None
-            index = int(index)
-    if index > len(items):
-        raise Exception("Index too high")
-    elif index == 0:
-        raise Exception("Lists are 1-indexed.")
-    # Negative nonzero indexes are already 1-indexed
-    # Positive non-zero indexes need correction
-    elif index > 0:
+        raise Exception("Invalid index. Must be a non-zero integer, 'random', or 'next'")
+
+        # Negative nonzero indexes are already 1-indexed
+        # Positive non-zero indexes need correction
+    if index > 0:
         index = index - 1
+
     # Need to return the index too as it could be random
     return items[index], index + 1 if index >= 0 else len(items) + index + 1
 
@@ -87,14 +106,22 @@ def get_list_size(list_name):
 
 
 def remove_from_list(list_name, index):
+    """Index is 1-indexed"""
     if not isinstance(index, int):
-        if not index.isdigit():
-            return
+        if not represents_int(index):
+            raise Exception("Invalid index")
         index = int(index)
+
+    found_list = db.session.query(List).filter_by(name=list_name).one_or_none()
+    if not found_list:
+        raise Exception("List not found")
 
     found_list_item, index = get_list_item(list_name, index)
     found_list_item_value = found_list_item.item
     db.session.delete(found_list_item)
+    # current_index is 0-indexed
+    if index - 1 <= found_list.current_index:
+        found_list.current_index -= 1
     db.session.commit()
 
     return found_list_item_value, index
